#153 Move tagging features to core.tagging module
