+++ pandas_dataclasses/core/__init__.py
@@ -1,6 +1,7 @@
-__all__ = ["api", "specs", "typing"]
+__all__ = ["api", "specs", "tagging", "typing"]
 
 
 from . import api
 from . import specs
+from . import tagging
 from . import typing
